services/requirement_gathering/pyproject.toml
[project]
name = "sdlc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "aiofiles>=24.1.0",
    "pydantic>=2.10.0",
    "sqlalchemy>=2.0.0",
    "openai-whisper>=20231117",
    "langchain>=0.3.9",
    "langchain-aws>=0.2.0",
    "langchain-openai>=0.2.0",
    "langchain-text-splitters>=0.3.0",
    "boto3>=1.35.0",
    "pypdf2>=3.0.0",
    "python-docx>=1.1.0",
    "python-multipart>=0.0.9",
]
